Add ‘mvd_part_filter’ cvar for muting control.
Setting this to positive value will enable filtering of MVD spectator
join and leave messages. If the spectator has not talked on a channel
during specified amount of time (in seconds), messages will not be
shown. Default value of zero means filtering is disabled, negative
values surpress messages for all clients.