Add ‘sys_forcegamelib’ cvar.
Attempt to clean up SV_InitGameProgs by removing hacky #ifdef stuff in
favor of using single variable specifying the absolute path to the game
library that should be tried first. It is empty by default on all
platforms. That means ‘.\release\gamex86.dll’ will no longer be tried by
default on Win32, and on non-Win32 platforms root of the libdir will not
be tried also. This needs to be explicitly requested on the command line.