allow logging data from a dumb tty on Unix
Continue emitting to stdout if EOF is hit on stdin. This helps running
the dedicated server from nohup(1), systemd and the like.

Please note that setting `sys_console=1' on the command-line is
obligatory in the event that a tty isn't attached. Launching without a
tty will default to `sys_console=0' in which case nothing is read or
emitted.

  % ./q2proded +set sys_console 1 +exec server.cfg

v2: [backed out]

v3:

- Disable polling stdin if EOF is hit (suggested by @skullernet).
- Restore original terminal settings on EOF (@skullernet)

Thanks to @fishxz for additional testing.

Issue: #168