Align RT shader group handles on reported base alignment.
On my RTX 2070 SUPER HandleSize is 32, while BaseAlignment is 64.
It still works on this driver, but it is a spec error to use
unaligned shader binding offsets in CmdTraceRays.

Unfortunately this means we have to expand the handles out into
an aligned table, as far as I could tell from the usage.