Change fake emissive generation
Previously, the algorithn was to simply extract pixels above a certain dynamic brightness.
Some scaling and blurring was applied as well.
The results were usable but could be crude-looking.

The new approach determines all 'bright' pixels (largest component above a certain) and
applies some blurring as well as some filtering with the original image.
After some experimentation with the details this resulted in an algorithm that
seems to extract emissive images that seem to be relatively faithful to the appearance
of the "light" part of original textures.