Explicitly track active renderer instead of relying on vid_rtx.
A lot of code assumes the value of vid_rtx also reflects the current renderer.
However, it can happen that this is not the case, causing code to rely on this
assumption to break (see #198).

Address by explicitly storing the currently active renderer and updating that
field on renderer initialization/shutdown. Also change any code that used vid_rtx
to check the renderer to use that new value.