Use exact number of cluster lights when assembling light lists in inject_model_lights()
Previously, space was made in the light list offsets that amounted to the maximum
number of lights in a cluster, seen over the lifetime of the map. Unfortunately,
this affect light sampling after a large spike of dynamic lights in a cluster,
possibly resulting in a more 'noisy' appearance - as described in #226.

However, the exact number of lights in a cluster was actually already counted
and can be used to tightly pack the light lists, avoiding the permanent effect
on sampling after a spike of dynamic lights.