Vastly reduce LIGHT_COUNT_HISTORY value
That constant, introduced in #234, is a pretty conservative value,
but unnecessarily large number.

The smallest number I found to work well is three - which makes sense,
if you think about it: ASVGF only needs "history" from the previous frame,
so we need: one for previous frame rendered, one for current frame rendering,
and one more for upload.